refactor: 패키지 구조 변경
(cherry picked from commit 7a25a603b6e0f2067c8177aa70893e946a4e773d)
