+++ src/main/java/webserver/http/Session.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http;
 
 import java.util.HashMap;
 import java.util.Map;
