+++ src/main/java/webserver/http/SessionManager.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http;
 
 import java.util.HashMap;
 import java.util.Map;
